requirements.txt
dearpygui
flask
kivy
pywebview